Updates to move writing to common routing_deletions and get routine data deletion working
