+++ airflow/dags/jper_scheduler/routing_deletions.py
@@ -1,23 +1,260 @@
 import os
-from posix import stat
+import math
+import json
+import fcntl
 import shutil
 from datetime import datetime, timezone
 from pathlib import Path
 
+import esprit
 from service.models.routing_history import RoutingHistory
 from jper_scheduler.publisher_transfer import PublisherFiles
 from octopus.core import app
 from octopus.modules.store import store
 
 from service import models
-from jper_scheduler.utils import create_routing_history_record
+from jper_scheduler.utils import get_notifications_for
 
 dryRun = app.config.get("AIRFLOW_DELETION_DRY_RUN", True)
+del_log_path = app.config.get("AIRFLOW_DELETION_LOGS_PATH", '/logs/data_deletion_logs')
+notifications_to_process = app.config.get("AIRFLOW_DELETION_NOTIFICATION_BATCH_SIZE", 5000) # Notifications to process at a given time.
 
+# Elasticsearch configuration
+max_query = app.config.get("AIRFLOW_DELETION_MAX_QUERY", 2000) # Max number of notifications to fetch in one query from ES - adjust as needed based on performance and memory constraints.
+host = app.config.get("ELASTIC_SEARCH_HOST", "localhost")  # includes port
+port = host.split(':')[-1]
+host_name = host.split(port)[0][:-1]
 
-# This class inherits from PublisherFiles (publisher_transfer.py) and will only
-# perform deletions.
-# Hopefully this will keep the publisher_transfer.py clean.
+##### Useful stand-alone functions
+
+def find_notifications_from_routing_history(since, upto, publisher_id, status_values, rerouting, deletion_reason):
+    a = RoutingHistory()
+    records = a.pull_records(since=since, upto=upto, status="error", publisher_id=publisher_id)
+    if not records or len(records) == 0:
+        app.logger.info("No records returned from open search matching query.")
+        return []
+    num_records = records.get('hits', {}).get('total', {}).get('value', 0)
+    if num_records == 0:
+        app.logger.info("No records returned from open search matching query")
+        return []
+
+    info_to_run = []
+    for hit in records.get('hits', {}).get('hits', []):
+        for note_state in hit.get("notification_states", []):
+            notification_id = note_state.get("notification_id", "None")
+            if "scheduled" in deletion_reason:
+                info_to_run.append((notification_id, status_values, rerouting, deletion_reason, publisher_id, hit.get("id")))
+            else:
+                info_to_run.append((notification_id, status_values, rerouting, deletion_reason, publisher_id))
+    return info_to_run
+
+#####
+
+def find_notifications_from_ES_directly(conn, since, upto, publisher_id, status_values, rerouting, deletion_reason):
+    page_size = 1000
+    records = get_notifications_for(
+        conn=conn,
+        since=since,
+        upto=upto,
+        page=1,
+        page_size=page_size,
+        publisher_id=publisher_id,
+    )
+    scroll_id = records["_scroll_id"]
+    if not records or len(records) == 0:
+        app.logger.error("Open search returned null record- exiting")
+        return []
+    num_records = records.get("hits", {}).get("total", {}).get("value", 0)
+    if num_records == 0:
+        app.logger.info("No records returned from open search matching query - exiting")
+        return []
+
+    info_to_run = []
+    for hit in records["hits"]["hits"]:
+        notification_id = hit["_id"]
+        info_to_run.append((notification_id, status_values, rerouting, deletion_reason, publisher_id))
+
+    if num_records > page_size:
+        page = 2
+        num_pages = int(math.ceil(num_records / page_size))
+        for page in range(2, 1 + num_pages):
+            records = get_notifications_for(
+                conn=conn,
+                since=since,
+                upto=upto,
+                scroll_id=scroll_id,
+                page=page,
+                page_size=page_size,
+                publisher_id=publisher_id,
+            )
+            if not records or len(records) == 0:
+                app.logger.info(f"Open search returned null record for page {page} - finishing")
+                break
+            for hit in records["hits"]["hits"]:
+                notification_id = hit["_id"]
+                info_to_run.append((notification_id, status_values, rerouting, deletion_reason, publisher_id))
+    return info_to_run
+
+##### Write to a file, every notification that we have been asked to delete.
+def write_notifications_to_delete(params, del_file, info_to_run=None):
+    if os.path.exists(del_file):
+        # File exists! Should not happen?
+        app.logger.warning(f"File {del_file} already exists - should not happen...")
+        del_file = del_file.with_name(f"{del_file.name}.bak")
+        if os.path.exists(del_file):
+            raise FileExistsError(f"Original and Backup files {del_file} already exists - should not happen...")
+    notification_id = params.get("notification_id", None)
+    publisher_id = params.get("publisher_id", None)
+    publisher_email = params.get("publisher_email", None)
+    status_values = params.get("status_values", [])
+    upto = params.get("upto", None)
+    brom = params.get("from", None)
+    rerouting = params.get("rerouting", None)
+    deletion_reason = params.get("deletion_reason", None)
+
+    app.logger.info("Parameters to search for routing history records:")
+    app.logger.info(f"publisher_id: {publisher_id}")
+    app.logger.info(f"publisher_email: {publisher_email}")
+    app.logger.info(f"status_values: {status_values}")
+    app.logger.info(f"from: {brom}")
+    app.logger.info(f"upto: {upto}")
+
+    if not info_to_run:
+        if notification_id:
+            info_to_run = [notification_id]
+        elif status_values[0] == 'failure': # Error
+            info_to_run = find_notifications_from_routing_history(brom, upto, publisher_id, status_values, rerouting, deletion_reason)
+        else:
+            if len(status_values) == 2:
+                index = "jper-routed*,jper-failed"
+            elif status_values[0] == "success-routed":
+                index = "jper-routed*"
+            elif status_values[0] == "success-no-matches":
+                index = "jper-failed"
+            else: # Should never happen
+                app.logger.error(f"Unexpected status value(s): {status_values}")
+                return []
+            conn = esprit.raw.Connection(host_name, index, port=port)
+            info_to_run = find_notifications_from_ES_directly(conn, brom, upto, publisher_id, status_values, rerouting, deletion_reason)
+
+    app.logger.info(f"Total number of notifications to process: {len(info_to_run)}")
+    app.logger.info(f"Writing notifications to file {del_file}")
+
+    info_to_write = {}
+    info_to_write["publisher_id"] = publisher_id
+    info_to_write["publisher_email"] = publisher_email
+    info_to_write["status_values"] = status_values
+    info_to_write["from"] = brom
+    info_to_write["upto"] = upto
+    info_to_write["rerouting"] = rerouting
+    info_to_write["deletion_reason"] = deletion_reason
+    info_to_write["total_notifications"] = len(info_to_run)
+    info_to_write["remaining_notifications"] = len(info_to_run)
+    info_to_write["notifications"] = info_to_run
+
+    if not os.path.exists(del_file.parent):
+        os.makedirs(del_file.parent)
+    with open(del_file, 'a') as f:
+        f.write(json.dumps(info_to_write) + '\n')
+
+##### Return a list of notifications to from everything that is waiting
+
+def read_notifications_to_delete(max_map_length):
+    info_to_run = []
+    path = Path(del_log_path).rglob('TODO/*.json')
+
+    kount = 0
+    done = False
+    for file in path:
+        if not os.path.exists(file):
+            app.logger.info(f"File not found: {file}. Should not have happened, skipping ...")
+            continue
+        with open(file, 'r') as f:
+            data = json.loads(f.read())
+        for notification in data["notifications"]:
+            notification.append(str(file))
+            info_to_run.append(notification)
+            kount += 1
+            if kount >= notifications_to_process or kount >= max_map_length:
+                done = True
+                break
+        if done:
+            break
+    app.logger.info(f"Found {kount} notifications to delete")
+    return info_to_run
+
+##### Update the deletion log files. Of course, we need to read the file, update the buffer, and write it back.
+
+def update_deletion_log_files(del_log_file, airflow_log_url, notification_id, status, doi):
+    if not del_log_file:
+        app.logger.info(f"No deletion log file specified : {del_log_file}. Returning.")
+        return
+    if "TODO" not in del_log_file.parts:
+        app.logger.info(f"Deletion log file {del_log_file} is not a TODO file. Returning.")
+        return
+
+    words = []
+    for word in del_log_file.parts:
+        if word == "TODO":
+            if status == "success":
+                word = "DONE"
+            else:
+                word = "FAILED"
+        words.append(word)
+    final_file = Path("/".join(words)[1:])
+    if not os.path.exists(final_file.parent):
+        os.makedirs(final_file.parent)
+
+    # Read the input log file, update it, and write it back
+    # Note that we will only update with a single notification for a single task
+    tmp_list = []
+    with open(del_log_file, 'r+') as f:
+        fcntl.flock(f, fcntl.LOCK_EX)
+        data = json.loads(f.read())
+
+        for note_list in data["notifications"]:
+            if notification_id in note_list:
+                tmp_list = note_list
+        try:
+            data["notifications"].remove(tmp_list)
+        except ValueError:
+            app.logger.info(f"No notification found with id {notification_id} in DONE log file")
+            app.logger.info(f"You are likely rerunning a successful task. Stopping here.")
+            raise ValueError(f"No notification found with id {notification_id} while updating DONE log file")
+        data["remaining_notifications"] = len(data["notifications"])
+        f.seek(0)
+        f.truncate(0)
+        f.write(json.dumps(data))
+        fcntl.flock(f, fcntl.LOCK_UN)
+    if data["remaining_notifications"] == 0:
+        del_log_file.unlink()
+
+    # Read the updated log file and append the new notification that has just been deleted
+    if not tmp_list:
+        app.logger.info(f"No notification found with id {notification_id}?")
+        raise ValueError(f"No notification found with id {notification_id} while updating DONE log file")
+    tmp_list.extend([doi, airflow_log_url])
+    if final_file.exists():
+        with open(final_file, 'r+') as f:
+            fcntl.flock(f, fcntl.LOCK_EX)
+            data = json.loads(f.read())
+            data["notifications"].append(tmp_list)
+            data["completed_notifications"] = len(data["notifications"])
+            f.seek(0)  # Go to the beginning of the file
+            f.truncate(0)  # Clear the file before writing for safety
+            f.write(json.dumps(data))
+            fcntl.flock(f, fcntl.LOCK_UN)
+    else:
+        data["completed_notifications"] = 1
+        data["notifications"] = [tmp_list]
+        # The rest of the data comes from reading the above file
+        with open(final_file, 'w') as f:
+            fcntl.flock(f, fcntl.LOCK_EX)
+            f.write(json.dumps(data))
+            fcntl.flock(f, fcntl.LOCK_UN)
+
+##### ##### #####
+# This class inherits from PublisherFiles (publisher_transfer.py) and will only perform deletions.
 class RoutingDeletion(PublisherFiles):
     def __init__(self, publisher_id=None, routing_id=None):
         if not publisher_id or not routing_id:
