+++ local.template.cfg
@@ -240,7 +240,8 @@ AIRFLOW_REPROCESS_NOTIFICATION_BATCH_SIZE = 1000
 
 AIRFLOW_DELETION_DRY_RUN = True
 AIRFLOW_DELETION_MAX_QUERY = 5000
-AIRFLOW_DELETION_LOGS_PATH = '/logs/data_deletion_logs'
+AIRFLOW_DELETION_LOGS_PATH = '/logs/request_deletion_logs'
 AIRFLOW_DELETION_NOTIFICATION_BATCH_SIZE = 3000
 
 AIRFLOW_DELETE_OLD_SCHED = None
+AIRFLOW_DELETION_LOGS_DAILY_PATH = '/logs/daily_deletion_logs'
